+++ changelog.md
@@ -2,3 +2,4 @@
 
 ## devel
 
+* add current logs for all mapi container pods